refactor: isolate runtime prompt helpers
